Organize background subtraction packages into one single package
Organize background subtraction related packages into one single package to make them clean.
