+++ extract_aris/bg_sub/BBoxManagerIcon.ico
