+++ extract_aris/bg_sub/test_assets/test_json/0.png
